VDB-5757 v0-decompress-local.h

+++ libs/sraxf/v0-decompress-local.h
@@ -230,7 +230,7 @@ typedef struct SRALocal
         int _bits_done = 0;						\
         int _nbits = nbits;						\
                                                                         \
-        if (_bitpos + _nbits + (_shift << 3) > max_bits) {              \
+        if (_bitpos + _nbits + (_shift << 3) > (unsigned int)max_bits) {\
             fprintf(stderr, "legacy code trying to read %u bits beyond end of data at %s, line %d.\n", _bitpos + _nbits + (_shift << 3) - max_bits, __FILE__, __LINE__); \
             _nbits = max_bits - (_bitpos + (_shift << 3));              \
         }                                                               \
@@ -282,7 +282,7 @@ typedef struct SRALocal
         int _bits_done = 0;						\
         int _nbits = nbits;						\
                                                                         \
-        if (_bitpos + _nbits + (_shift << 3) > max_bits) {              \
+        if (_bitpos + _nbits + (_shift << 3) > (unsigned int)max_bits) {\
             _nbits = max_bits - (_bitpos + (_shift << 3));              \
         }                                                               \
         while (_nbits > 0) {                                            \
